refactor(general_functions): migrate get_recid method to requests_rdm
Migrate get_recid method to requests_rdm because it is a more fitting
module for it. Also set the references to it accordingly.
