+++ invenio_rdm_pure/source/rdm/run/changes.py
@@ -24,7 +24,7 @@ from ...pure.requests_pure import (
 from ...reports import Reports
 from ..add_record import RdmAddRecord
 from ..delete_record import Delete
-from ..general_functions import GeneralFunctions
+from ..requests_rdm import Requests
 
 
 class PureChanges:
@@ -35,7 +35,7 @@ class PureChanges:
         self.add_record = RdmAddRecord()
         self.report = Reports()
         self.delete = Delete()
-        self.general_functions = GeneralFunctions()
+        self.rdm_requests = Requests()
 
     def get_pure_changes(self):
         """Gets from Pure 'changes' endpoint all records that have been created / updated / deleted.
@@ -152,7 +152,7 @@ class PureChanges:
             self.report.add(report)
 
             # Gets the record recid
-            recid = self.general_functions.get_recid(uuid, self.global_counters)
+            recid = self.rdm_requests.get_recid(uuid, self.global_counters)
 
             if recid:
                 # Deletes the record from RDM
