+++ invenio_rdm_pure/source/rdm/versioning.py
@@ -11,7 +11,6 @@ import json
 
 from ..general_functions_source import add_spaces
 from ..reports import Reports
-from .general_functions import GeneralFunctions
 from .record_manager import RecordManager
 from .requests_rdm import Requests
 
@@ -23,7 +22,6 @@ class Versioning:
         """Description."""
         self.report = Reports()
         self.rdm_requests = Requests()
-        self.general_functions = GeneralFunctions()
 
     def get_uuid_version(self, uuid):
         """Gives the version to use for a new record and old versions of the same uuid."""
